bmi_calculator.py

+++ bmi_calculator.py
@@ -2,7 +2,7 @@ height = input("enter your height in m: ")
 weight = input("enter your weight in kg: ")
 
 new_height= float(height)
-new_weight= float(weight)
+new_weight= int(weight)
 bmi= (new_weight / (new_height ** 2))
 new_bmi=int(bmi)
 print("Your bmi is :", new_bmi)
